+++ tests/test_converters/test_flatten.py
@@ -6,6 +6,7 @@ from returns.context import (
     RequiresContextResult,
 )
 from returns.converters import flatten
+from returns.future import Future, FutureResult
 from returns.io import IO, IOFailure, IOSuccess
 from returns.maybe import Some
 from returns.result import Failure, Success
@@ -23,7 +24,9 @@ from returns.result import Failure, Success
 
     # Nope:
     (Failure(Failure('a')), Failure(Failure('a'))),
+    (Failure(Success('a')), Failure(Success('a'))),
     (IOFailure(IOFailure('a')), IOFailure(IOFailure('a'))),
+    (IOFailure(IOSuccess('a')), IOFailure(IOSuccess('a'))),
 ])
 def test_flatten(container, merged):
     """Ensures that `flatten` is always returning the correct type."""
@@ -55,3 +58,41 @@ def test_flatten(container, merged):
 def test_flatten_context(container, merged):
     """Ensures that `flatten` is always returning the correct type."""
     assert flatten(container)(...) == merged(...)
+
+
+@pytest.mark.anyio
+async def test_flatten_future(subtests):
+    """Ensures that `flatten` is always returning the correct type."""
+    futures = [
+        # Flattens:
+        (Future.from_value(Future.from_value(1)), Future.from_value(1)),
+        (
+            FutureResult.from_success(FutureResult.from_success(1)),
+            FutureResult.from_success(1),
+        ),
+    ]
+
+    for container, merged in futures:
+        with subtests.test(container=container, merged=merged):
+            assert await flatten(container) == await merged
+
+
+@pytest.mark.anyio
+async def test_non_flatten_future(subtests):
+    """Ensures that `flatten` is always returning the correct type."""
+    futures = [
+        # Not flattens:
+        (
+            FutureResult.from_failure(FutureResult.from_failure(1)),
+            FutureResult.from_failure(FutureResult.from_failure(1)),
+        ),
+    ]
+
+    for container, merged in futures:
+        with subtests.test(container=container, merged=merged):
+            assert isinstance(
+                (
+                    await flatten(container)  # noqa: WPS437
+                ).failure()._inner_value,
+                FutureResult,
+            )